requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Anthropic Claude API
anthropic==0.39.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# Redis
redis==5.0.1
celery==5.3.6

# Document Processing
pypdf2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
pytesseract==0.3.10

# ML/Embeddings
sentence-transformers==2.3.1
numpy==1.26.3

# Storage
boto3==1.34.34
minio==7.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.26.0
aiofiles==23.2.1
psutil==5.9.8
